+++ src/components/layout/header/components/TopLogo.js
@@ -1,7 +1,7 @@
 function TopLogo() {
     return (
         <div className='top-logo'>
-            <img src='/logo-toss-blue.png' />
+            <img src='/logo-toss-blue.png' alt="말랑 로고" />
         </div>
     );
 }
